Check that $CONFIG_MK is writable.
I discovered this issue having installed XCode from the command line. When
running config, it would ask me to run as root so that I could accept the
license agreement. Running as root creates a $CONFIG_MK, with root priveleges.
It is only then that I was told that I should run a special tool instead to
accept the license agreement. Running config afterwards, clearly disallowed me
to overwrite the config, but the script didn't report this in a pretty way.