test-cover mode
This makes use of the already existing check-cover-norun Makefile target. The
coverage is generated both for the language binding source and core. The first
allows to measure how much of the binding is tested, the last (together with
the first) allows to measure how much of the core functionality is implemented
and tested.