Fix memory leak when breaking out of a for-in loop
Use an autoreleased NSMutableArray to keep the temporary objects alive rather
than a heap-allocated id[] so that the last block of objects will get
autoreleased if the iteration does not complete. This results in iteration
breaking if the user drains the autorelease pool while iterating over the
RLMArray, but that's no longer supported with ARC and was always a bad idea.