Fix some issues with multilevel link queries
table->link() mutates the table rather than returning a new thing, which
resulted in queries which did not consume the link chain breaking later queries
on the same table.

Searching for "ANY array.col = object" doesn't actually work; it appeared to
work in the tests because was actually searching `array` for an entry with the
same row number as `object`, which coincidentally was the same as the
containing object's row number.