Avoid holding a strong reference to RLMRealms for notifications
If the thread's runloop isn't actively running when the notification is sent,
the notification will keep the RLMRealm alive until the thread's runloop is run
again, since performSelector:onThread:withObject:waitUntilDone: keeps the
receiver alive. Work around this by using a separate object which holds a weak
reference to the realm. Should fix the crashes in the background refresh test.