Make the core version check case-insensitive
It turns out that newer versions of pandoc have started capitalizing headers
when converting markdown to text, which breaks the version check for local core
builds.

It would be nice to be getting the version number from something other than a
text file generated by a third-party tool with not guarantees about its output
format.