Rewrite the Swift class support in Objective-C
Turns out that you can't use Swift in a framework on iOS 7, so move everything
but the convenience functions in Objective-C and switch to building a static
framework. The remaining Swift code is left in a single file which must be
included directly in the user's application (if they want to use it).

This happens to add support for optional RLMObject properties in Swift.