Add the release packaging scripts to build.sh
This keeps the same release flow of Jenkins assembling tested artifacts into
the final release package, and just moves the scripts for doing so into
build.sh so that changes to the release packaging can be done via PRs. It also
adds a test target which mimics the CI setup's assembly flow (but not any of
the actual releasing parts) so that changes to the release tasks can be tested
locally.