Get the version in a way that actually works with Xcode 5
Getting the version number from the plist only works when Realm is built as a
dynamic framework, as otherwise our plist doesn't actually end up in the final
executable at all. As our release builds are currently built with Xcode 5 and
thus not dynamic frameworks, this resulted in Realm reporting that an update
was available on every check.