Some issue with a change in timing caused the realm to load after the windowcontroller.
Now this has been handled, but it is still not ideal. A perfect solution will likely not be possible until we get rid of the document based approach, which I hope to do next week, or late this week, as it complicates everything having to do with loading and opening files and brings no obvious benefit.