Fix a crash in CFRunLoopSourceInvalidate
It turns out that the documentation's recommendation to release run loop
sources immediate after adding them to the run loop is a rather bad idea, as
the run loop releases its sources when it's shut down, so we'd crash if the
kqueue waiter outlived the source thread.

The test forces this to happen by holding on to the RLMRealm after the thread
dies, but it can happen inconsistently without that.

Closes #1608.