Avoid using private implementation details in the Swift tests
Explicitly wrap things creating Realms in `setUp` methods in `autoreleasepool`
blocks to avoid the need to wrap all calls to `setUp` explicitly, and then just
call `super.invokeTest()` in an autorelease pool rather than trying to
replicate all of the work it does. Requires an obj-c helper to work around not
being able to use `super` in a block in swift 1.1.