Have RLMNotifier create a dedicated thread to listen for notifications rather than doing so on a global concurrent queue.
There is a relatively low cap on the number of worker threads that can
service the global concurrent queues. Using a dedicated thread to listen
for notifications avoids tying up a worker thread for the lifetime of
each RLMRealm instance, which prevents starving other GCD clients when a
large number of RLMRealm instances are live.