Fix running the Swift tests on iOS devices.
My changes in #2008 broke them. Xcode requires that TEST_HOST be set for
iOS device tests, but the conditional way in which it was being set in
Tests.xcconfig was resulting in Xcode thinking it was unset.

To address this TEST_HOST is now set unconditionallly. This requires
some minor changes to allow TestHost.app to build and run on OS X.