Fix tests with the current master version of core.
Table::column<T> now checks that T matches the stored type of the
column. This breaks with the existing implementation of
ColumnOfTypeHelper that uses Table::column<Int> for DateTime
columns. This was done to avoid compilation errors from
query_expression.hpp that occurred due to DateTime's operator int64_t
being private. In core master, it's no longer private so our hack is no
longer necessary. As a temporary measure we conditionally select between
DateTime and Int for the column types to allow building with both the
latest core release and the current core master.