Clean up lingering in-memory realms for perf tests
If the test application crashes while running a performance test, the in-memory
realm file isn't deleted, and then is opened the next time the tests are run.
This works out fine when it's the same file format version, but upgrading
in-memory realms doesn't work correctly (and of course downgrading isn't
supported at all).