Support using initWithValue: in Swift property defaults
If an object has a property linking to another object and that property is
initialized using `initWithValue:`, that initializer may get run during schema
initialization before the linked class's schema is initialized and thus throw
an exception. We don't actually use any of the property values of linked
objects during schema init, so just do nothing in `initWithValue:` if the
class's schema has not been initialized.