Consolidate parameter validation in RLMRealmConfiguration
Previously some things were validated on property set, some were validated in
RLMRealm, and in RealmSwift some were validated on both. This makes
RLMRealmConfiguration's setters validate everything that it can,
Realm.Configuration validate nothing until it's converted to an
RLMRealmConfiguration, and RLMRealm trusts that the passed-in config is valid.

This makes supporting the objectstore config object a bit easier and makes the
tests more coherent.