Create a dynamicList method on Object
This method was primarily motivated by the need to get a list off
an Object with an unknown schema at runtime, because KVC
returns and will only accept a cast to the exact List<T: Object> type.

Users can unsafeBitCast a List from KVC to get the same behavior, but
it's nice to hide unsafeBitCasts away when they're required.