Fix and re-enable the multiprocess tests.
Now that a test host is used for all tests, it's not possible to run a
single test by reinvoking the application with a command-line
argument. Switch instead to locating `xctest` on the `PATH`, and
invoking it for use as the subprocess. This tweaking the subprocess's
creation of paths to Realms to ensure that they use the same directory
as the parent.