Fixed incorrect method type encoding for setters and getters.
The correct encoding for getters is "C@:", where C is the return type, @ is the receiver (self) and : the selector's first parameter (in this case there are no parameters).

The correct encoding for setters is "v@:C", where v is the return type, @ is the receiver (self), : the selector's first parameter, and C the type encoding of the first parameter.