Fix running tests on iOS devices with the dynamic framework.
Have Test.app embed Realm.framework in its Frameworks directory. Doing
this requires introducing a separate `TestHost iOS static` target for
the static framework tests since there is nothing for them to
embed. This target is identical to the existing `TestHost` target,
except for the lack of Embed Frameworks step.