[scripts] Initialize REALM_SWIFT_VERSION from branch
This ensures that on these branches, always the right source version is used instead of automatically determining it further down the script, which fails if the wrong Xcode version is selected globally via `xcode-select`.
This could be done explicitly on each branch, but that would cause that these branches diverge from master through a human-initiated action vs. just automatic commits by CI and could be easily forgotten for new such branches in the future.
Fixes #2941.