[RLMResults] Improve error thrown on missing argument of KVC collection operators
Improve the exception thrown if the argument of a unary KVC collection operator, such as '@sum.<keyPath>' is missing completely.
Before e.g. when only trying to access '@sum', it would have raised that as an unsupported operator, which would be misleading. With this change it will correctly report that as a missing argument.