Fix using -[RLMResults setValue:forKey:] on a key used in the query
This was broken by aadd90501c70fc09e17d04e5f25fd7b732e6a9cc, which indirectly
made it so that indexInSource: brings the Results back in to sync. Changing
that back would be difficult, so just directly iterate over a frozen tableview
instead.

Also adds a test to ensure this usage will continue to work in the future.

Fixes #3028.