enable PR CI builds on Travis CI
This will serve as a backup in case our main CI setup is experiencing
difficulties.

It will also allow non-Realm employees to see the results of some

I've only ported a subset of our standard CI PR matrix here because:

  1. Travis limits their matrix builds to 50 minutes *in aggregate*
  2. Some of our standard CI PR matrix jobs require dependencies not available
     to Travis, like codesigning certificates and provisioning profiles.

The downside to integrating this is that we'll have yet another file to update
when making changes to our build system or CI setup.

This commit also avoids grepping on RealmSwift/SwiftVersion.swift when the file
doesn't exist, preventing a warning from being printed.