Fix race conditions when an RLMRealm is deallocated from the wrong thread
The runloop notification source needs to hold a strong reference while the
block is running and a zeroing weak reference the rest of the time to handle
`CFRunLoopSourceInvalidate()` being called from a different thread.

Similarly, RLMNotificationHelper needs a zeroing weak reference to the RLMRealm
because the SharedRealm can outlive the RLMRealm and then crash when it tries
to deliver a notification to the RLMRealm. The reads of the weak pointer are in
autoreleasepools because ARC autoreleases pointers when acquiring a strong
reference to a weak pointer.

Fixes #3013.