Implement `+automaticallyNotifiesObserversForKey:` on `RLMObjectBase`.
Prior to this `RLMAccessorClassForObjectClass` was add implementations
of `+automaticallyNotifiesObserversOf<Property>` to the class for each
property in the schema. Given that the results of these methods is
cached by Foundation after the first call it's not necessary to optimize
them in this manner. In fact, doing so makes the creation of accessor
classes during schema initialization much slower due to the increased
number of calls to `class_addMethod`.

Fixes #3111.