Fix warnings when building via CocoaPods
Only include headers needed for the Realm.Private module in
private_header_files and just preserve_path the assembled include directory.
This makes it not copy a whole bunch of files into the framework which don't
need to be copied, and in some cases were generating warnings due to headers
with the same file name at different paths.

This requires fixing all of the objectstore include paths to not rely on
Xcode's header mapping stuff, which needed to be done anyway.

Also put the core headers in a different directory from the obj-c headers to
remove the need to do convoluted things to handle case-sensitive filesystems.