Make partial schema initialization more robust
Make it so that creating a standalone object only initializing the entire
shared schema if it needs to in order to validate link/array properties, and
otherwise only initialize that class's schema.

Fix a case where triggering initialization of the shared schema during partial
schema init would result in the classes used in the partial schema not
appearing in the shared schema.

Fix creating standalone instances of subclasses when the parent class has been
used in a partial schema and the shared schema has not been initialized.

Fix some probably innocuous data races.

Reduce the number of calls to class_replaceMethod() during schema init.