remove '_addNotificationBlock' from RealmCollectionType requirements
replacing it with the more accurate 'addNotificationBlock' vending Self as the
block parameter. This is a breaking change since it requires List's notification
block to vend '(List<T>?, NSError?)' rather than 'List<T>'. This change should
be evaluated to determine if it's the best course of action and if there are any
preferable alternatives.