[Migration] Keep old value when nullable to required auto migration
If change the property attribute from nullable to required, automatic migration works. But the values will be cleared. This behavior is difficult to understand what happened.
So keep the value if just changing nullability.

If old value is null, the new value will be an empty string, a zero, a zero length data or an unix epoch date.