Remove `addNotificationBlock(_:)` from `RealmCollectionType`.
It no longer compiles now that `Results` is not final. It was added as
part of the fine-grained notifications work and has not yet shipped in a
release, so it's ok to remove. We can look into bringing it back in the
future if we can find a way to do so.