Cache the default Realm file directory
This dramatically speeds up creating new Realm.Configuration objects with the
default initializer, as `RLMRealmPathForFile("default.realm")` took the bulk of
the time.

The main benefit of this is that it speeds up the Swift tests since we create a
very large number of Configuration objects, but it may also help user code that
does inefficient things.