Make Realm() less needlessly slow
When there was already a cached Realm instance on another thread nearly a third
of the run time was spent converting the default RLMRealmConfiguration to a
Realm.Configuration struct and then immediately back again. When there was a
cached instance for the current thread, this was nearly 100% of the run time.

Switch the testRealmCreationCached test over to testing opening the default
Realm rather than creating a new configuration object every time, as the latter
is hopefully not a realistic user scenario.