Remove RLMAccessorCode and objcType from RLMProperty
There was a bunch of code related to tracking the exact declared type of
properties in RLMProperty, but the only thing this was actually used for was
generating the accessor method overrides for int properties, and that code can
easily just read the required size itself.

Similarly, RLMAccessorCode was no longer used for much of anything.