Fix reading the shared schema from object instances
The overridden sharedSchema class method only worked properly when called
directly on the class it's overridden on, but actual object instances are
always a generated subclass of that class. This roundaboutly ended up working
by coincidence in most cases due to it finding the correct shared schema via
the lookup by name, but for KVO-generated subclasses with a different name it'd
instead end up trying to generate a new RLMObjectSchema for the runtime-created
subclass. For obj-c classes this worked and was just pointless extra work, but
for Swift classes it crashed.