Simplify Swift object init and subscripting logic a bit
Push more of the logic into obj-c to eliminate a bunch of wrappers and hopping
between layers for little benefit, and make it so that the Swift code has no
dependency on any private members of RLMProperty.

Don't do the eager init for DynamicObject as we can easily do it on access
(since the only access is through the subscript function under our control),
and remove the logic for optional numbers which was never actually used for
anything.