Rework query building to not use column numbers from RLMProperty
Explicitly pass in the Group to the query builder code and look up tables in it
and column numbers in the Tables rather than getting those indirectly from
RLMObjectSchema and RLMProperty. This eliminates any dependency on the private
members of the schema classes and makes the query code ready for runtime schema
changes.

Refactor the code by putting most of the free functions into a class that
stores the Schema, Query and Group to avoid having to pass around quite so many
values to every single function.