Fix `RLM_ERROR_ENUM` with Swift 2.3.
Swift 2.3 uses a version of clang for importing that understands the
`ns_error_domain` attribute without providing it any special semantic
meaning. Combined with the fact that Swift 2.3 applies special semantics
to enums declared by a macro named `CF_ENUM` or `NS_ENUM`, this was
resulting in the error code enums being imported in an undesirable way.

To work around this we delegate to `NS_ENUM` for the actual declaration
of the enum, inserting the `ns_error_domain` attribute as part of the
enum name. This results in warnings due to the attribute being ignored
in some places the name is used within the macro expansion, so we ignore
those warnings within the expanded code.

We should switch to using `NS_ERROR_ENUM` once it is available.