Introduce `asObjectiveCAccessor` method on `Object`
We expose the Realm Objective-C API operating on `RLMObject`, but many
of the methods can operate on any subclass of `RLMObjectBase`. As such, we
currently bitcast `Object` to `RLMObject` to call these methods. This
commit introduces an `asObjectiveCAccessor` function on `Object` to clarify
the intent of the cast, to prevent mistakes (such as bitcasting `Any` to
`RLMObject` without first force casting to `Object`), and to discourage
raw usage of `unsafeBitCast` in our codebase.