rework cocoa permission change implementation
instead of having a base Objective-C class, with the same PermissionChange
class name in both Realm Objective-C and Realm Swift, with a complex array
of @compatibility_aliases, type aliases and subclasses.

In its place, hardcode a mapping of Objective-C and Swift classes that should
use a common table name.

This is a temporary hack that should be replaced with a more generalized
approach to map classes to custom tables.