Support Swift files with overridden @objc names
Previously, Realm would assert when creating a schema for this object
because it only considered files with modular class names to be Swift
files.

```
@objc(Foo) class Foo: Object {}
```

This is problematic for projects that do all manipulation of an
instance in Swift but still need to pass it between Objective-C classes.

Now we check for a dot in the name or inheriting from RealmSwiftObject.