return class name from `Object.className` instance method/var
The intention when these were added was to prevent them from being
overridden in user models by marking these overrides as `final`, since users
were trying to change the name of the table in the Realm by returning a
different value than the default.

However, the `static` keyword was omitted by accident, causing
this to not work as intended at all.

So now that we need to keep this around to avoid a breaking change
to conform to SemVer, might as well have it return the correct values.