Change our approach to resetting iOS simulators
We used to do the following:
1. Kill any running simulator processes.
2. Shut down any running simulator devices.
3. Erase the contents of all simulator devices.
4. Launch Simulator.app.
5. Wait for Simulator.app to boot a simulator device.

Steps 3, 4 and 5 were somewhat problematic as there's no guarantee that
the simulator devices will exist and be in a valid state that will allow
Simulator.app to launch them. This often resulted in step 5, waiting for
the simulator devices to boot, timing out.

We now do the following:
1. Kill any running simulator processes.
2. Delete any unavailable simulator devices.
3. Shut down any running simulator devices.
4. Delete all remaining simulator devices.
5. Create a fresh simulator device for each valid platform / device type
   combination.