Don't delete unavailable simulator devices.
Unavailable simulator devices are those whose simulator runtime belongs
to a different version of Xcode. Deleting these devices means we leave
a different set of simulators configured at the end of the script
execution, which can cause problems for anyone that assumes the default
set of simulators will always be available.

It should not be problematic to leave these unavailable devices around
as we still kill any processes associated with them, and Xcode will
never attempt to use an unavailable simulator device.