Enable NSPredicate evaluation with RLMArray as subquery collection
When evaluating an NSPredicate on an RLMObject, an exception is thrown if the subquery's collection is an RLMArray instance although it should work just fine.

`-[NSPredicate evaluateWithObject:]` eventually calls `-[NSSubqueryExpression expressionValueWithObject:context:]` when
the predicate is a subquery. The problem is that this method is incorrectly implemented, it goes roughly like this:

```objc
id collection = [self.collection expressionValueWithObject:object context:context];
if (![collection isNSArray__] || ![collection isNSSet__] || ![collection isNSOrderedSet__])
{
    NSString *reason = @"Can't perform collection evaluate with non-collection object."
    @throw [NSException exceptionWithName:NSInternalInconsistencyException reason:reason userInfo:nil];
}

for (id element in collection)
{
    // goes on with evaluation
}
```

So, instead of testing if the collection conforms to the `NSFastEnumeration` protocol (which RLMArray does) the
collection is tested against three known Foundation collections.

Pretending that RLMArray is an NSArray by implementing the `isNSArray__` method enables NSPredicate evaluation with RLMArray as subquery collection.

Since the `isNSArray__` method is not public, some care was taken to construct the selector dynamically.