Work around a hang seen when building Realm using Xcode 9b5 with Carthage.
Xcode 9 beta 5 appears to hang when computing the build settings for the
PreloadedData scheme in the tvOS examples project. At the point of the
hang, `xcodebuild` is initializing some portion of the code signing
subsystem. The only thing interesting about the PreloadedData target is
that it has a development team specified, unlike all of the other
targets in our example projects. Removing the development team prevents
the hang from being hit.