Fix `sh build.sh download-sync` when the Xcode command-line tools are selected.
We use the output of `xcode-select -p` to determine which version of
Xcode we should be using. The Xcode command-line tools are not a
complete version of Xcode, so we must do something different if they are
the active Xcode version. We now emit a warning, find the most recent
version of Xcode.app, and use that instead.