Translate exceptions thrown when initializing RLMSyncManager to RLMExceptions
Currently if an error occurs (such as the user denying access to the keychain)
the program just crashes when the exception bubbles out of dispatch_once() with
no indication as to what went wrong. Rethrowing the error as a NSEception and
using std::call_once() (which allows propagating exceptions) there is at least
an error message printed.