Add support for overriding table and column names
Get _realmObjectName back into a working state and fix the things that never
actually worked with it (such as links and arrays), and add _realmColumnNames
to allow overriding the column names used.

This is functionality we need to have in the public API eventually (for the
same reasons as we need it internally), but for now it's internal-only to
reduce the potential complications.