Don't copy the RLMSchema every time a swift Realm is opened with explicit object types
Translating the Realm.Configuration to a RLMRealmConfiguration was making a
copy of the RLMSchema, which on top of being somewhat slow was breaking the
object accessor caching and thus leaking some memory each time.

This leaves the property as `copy` in case there's anything actually relying on
that and just bypasses the property setter when converting types.