Fix converting a RLMSyncConfiguration with partial sync enabled to a Swift SyncConfiguration
`-[RLMSyncConfiguration realmURL]` was returning the composed partial
sync URL rather than the reference URL. This resulted in an exception
about the presence of `__partial` in the URL when the
`SyncConfiguration` was later used to open a Realm.

To address this `-realmURL` now returns the reference Realm. A few
internal callers were updated to accommodate this. It seems unlikely
that any external users are relying on this behavior, and if they
are... partial sync is still in beta.