Fix warnings when building with Xcode 9.3 and Swift 4.
One overload of `flatMap` is being renamed to `compactMap`. To avoid
having to deal with this, the one case where we use this has been
replaced by separate calls to to `filter` and `map`, since it's not in a
performance-sensitive code path. Another instance has been replaced by a
call to `map` since it never actually returns `nil`.