Rework RealmOptional to work when it outlives the parent object
Storing a weak pointer to the parent object and performing accesses via that
relies on the parent object always outliving the RealmOptional, and it turns
out that that doesn't always happen even with completely innocuous code.
Instead, rework it to work more like RLMArray with separate managed/unmanaged
implementations and a strongly-owned ObjectStore backing object for the managed
version.