Rework how ROS is downloaded and launched for tests
Automatically download ROS as part of running the tests rather than requiring
it to be done manually, and add a check to make sure that the installed copy is
actually the correct version.

Switch to using a simple script to launch ROS rather than `ros start`. This
makes it possible to support testing ROS functionality which requires
customizing ROS, and simplifies adjusting the ROS settings when debugging
things.

Drop support for PROVIDING_OWN_ROS. It didn't actually work, and is unnecessary
(the same end result can be achieved by installing a dev version of ROS to
where the test script looks instead).