Fix object deletion handling within migrations
We defer object deletions done within a migration to until the migration
completes to maintain a stable enumeration order and to ensure that the old and
new objects given to the user actually correspond to each other. The logic for
this turned out to have a number of bugs, including sometimes failing to
exclude objects which were soft-deleted from the enumeration, deleting the
incorrect objects at the end of the migration, and just plain crashing.

The revised version is probably also significantly faster as it eliminates some
quadratic-time operations and avoids instantiating so many objects.