Fix a crash on non-@objc `String?` properties
 #5713 made it so that optional Swift properties of a supported data type (i.e.
`String?` and object links) which do not have ivars due to not being declared
as `@objc dynamic` would crash rather than merely fail to actually read and
update the managed value when the property is used.

The idea behavior here would be to throw an error on schema initialization as
it's an invalid model declaration, but that's a breaking change (as such
objects do work correctly when unmanaged), so for now this just restores the
pre-3.5.0 behavior.