Re-fix a crash on non-@objc `String?` properties
The previous fix in e320720e54482620ccc139d10f30583b3472b874 only worked if the
ivar happened to be `nil`. On watchOS 5 this is no longer always the case, and
so we would sometimes try to treat something that was a pointer to something
other than a RealmOptional as a RealmOptional.

This fixes it more robustly by instead just applying the workaround to *all*
non-@objc `String?` properties to avoid reading the ivar at all.