Add subscription management functions
This adds `-[RLMRealm subscriptions]` which returns a `RLMResults` of
subscriptions and `-[RLMRealm subscriptionWithName:]` to look up a specific
subscription, along with Swift wrappers for these. Rather than having a special
thing using LIKE syntax as JS has, the `RLMResults` is simply filterable like
any other collection.

The implementation of this is made awkward by that we already have a
RLMSyncSubscription type that's not an RLMObject and doesn't behave like an
RLMObject. Fortunately, obj-c is sufficiently dynamic that we can sometimes
return the existing objectstore `Subscription`-backed type and sometimes return
a new `RLMObjectBase`-backed type and have them behave (hopefully) identically.