Package a valid obj-c header for Swift 5
Swift 5 has changed how the generated header for fat libraries is assembled to
handle the case where it isn't identical on every architecture. However, this
doesn't automatically handle simulator+device fat libraries, since those aren't
officially supported. Since we use the header from the device build of the
library, this resulted in errors like "Cannot find interface declaration for
'RealmSwiftObject', superclass of 'UserDefinedSubclass'" when building for the
simulator.

Currently the generated header on every platform is identical so we can just
grab one of the platform-specific ones and package that instead of the
incompletely merged one. This will probably break at some point in the future,
so the build script verifies that this is a safe thing to be doing.

This doesn't touch Carthage packaging because Carthage should be releasing a
new version which deals with this problem any day now.

Note that we don't really support using RealmSwift from objective C and this is
mostly just to avoid causing problems with using other Swift libraries in the
same project.