Ensure we get permission objects when using asyncOpen on a query-based Realm
The server doesn't send us all of the permission objects until the first time
we create a subscription on the client. This causes problems when trying to
modify permissions immediately after using asyncOpen to open a Realm which
doesn't already exist on the client, as we expect permission objects to exist
for each of the classes.

Work around this by creating a dummy subscription (that'll never actually get
run) if the Realm we get from asyncOpen has zero subscriptions in it.