Rework managed object initialization
RLMObject had a large number of designated initializers that weren't really
intended to be designated initializers which have stuck around for years due to
it being a breaking change to remove them. This cuts them down to just `init`,
eliminating some confusing boilerplate that was needed for custom initializers
in Swift.