Rework and simplify release packaging build scripts
Get rid of all the stuff for building and packaging objc only and instead just
package the objc frameworks built as part of building the swift framework. Also
remove a bunch of old build targets that used to be used by Jenkins but hasn't
been ever since the jobs were split into smaller chunks that can be run in
parallel.