Restore support for @objc enums on Swift objects
This previously worked largely by coincidence due to these properties being
reported to the objc runtime as int properties. Shifting to Swift-based
introspection makes it possible for us to tell the difference, which means that
we need to explicitly handle them.

Due to `extension RawRepresentable: ManagedPropertyType { ... }` not being
legal Swift (you can't make an existing protocol refine another protocol in an
extension), this requires adding a `RealmEnum` protocol that enums have to
explicitly conform to. This also requires making ManagedPropertyType
underscore-prefixed and public rather than internal because RealmEnum needs to
refine it.

The upside of all this is that storing an enum in a RealmOptional now works,
when it didn't in <4.0.