Hide the spm installation example project from Carthage
As part of deciding what to build, Carthage calls `xcodebuild -list` on every
xcodeproj and xcworkspace in the repository. For projects which use SPM, this
performs package resolution and downloads all of the project's dependencies.
This would merely make things needlessly slow, but Carthage also imposes a one
minute timeout on this and with a slow internet connection it can take more
than a minute to clone the repositories.

Work around this by assembling the xcodeproj as part of running the test rather
than having it stored normally in the repo.