Fix compilation via Carthage with Xcode 12
Something that Carthage is doing makes xcode incorrectly think that it can skip
the ProcessXCFramework build step for the second scheme it builds even though
the intermediate directory where the xcframework files are copied to are
different for the two schemes. Work around this by touching each file in the
xcframework after the build, which effectively disables incremental
compilation.